requirements.txt
google-generativeai
pillow
cadquery
tqdm
pythonocc-core
trimesh
